requirements.txt
requests
pandas
duckdb
streamlit
python-dotenv
pytest
